Fix to replace main.swift with @main in Demo
